Correct value for `stripLengthMeters` parameter.
Changed the value from 5 to 2.5. Each reel of LEDs is 5m long, but in
this project each reel will be split into two strips: one extending from
the center of the frustum to the outer circle, and extendeding back. As
used in the program, the parameter `stripLengthMeters` is the slant height
of the frustum, so it should represent the length of half a reel of LEDs,
not a whole reel. Thus it should be 2.5, not 5.